random seed util;  int32_t seed -> int64_t seed; ofrecord_decoder seed impl; test pass of new_data_pre.py

+++ oneflow/customized/image/random_crop_generator.h
@@ -10,7 +10,7 @@ using AreaRange = std::pair<float, float>;
 
 class RandomCropGenerator {
  public:
-  RandomCropGenerator(AspectRatioRange aspect_ratio_range, AreaRange area_range, int32_t seed,
+  RandomCropGenerator(AspectRatioRange aspect_ratio_range, AreaRange area_range, int64_t seed,
                       int32_t num_attempts);
 
   void GenerateCropWindow(const Shape& shape, CropWindow* crop_window);
@@ -21,7 +21,7 @@ class RandomCropGenerator {
   std::uniform_real_distribution<float> aspect_ratio_log_dis_;
   std::uniform_real_distribution<float> area_dis_;
   std::mt19937 rand_gen_;
-  int32_t seed_;
+  int64_t seed_;
   int32_t num_attempts_;
 };
